Add FastClick compatibility
Fastclick does not seem to properly handle mousedown events. Adding the onclick event allows the addon to be used on mobile devices in combination with FastClick.

I'm trying to get this fixed in FastClick library, but the maintainer does not seem to be very active considering the many open issues / pull requests. For now this might be the best way to allow users to continue using this very cool addon.

Partly solves issue #248 described in ember-power-select "Incompatibility with fastclick?".
https://github.com/cibernox/ember-power-select/issues/248