Remove double binding functionality in the `opened` property.
Having this field being a special case was the source of some bugs
that had to do with the way the bindings were propagated. Bad shit.

All the component followed a strict DDAU policy except this field,
so I've removed it. You can still pass `initiallyOpened=true` to
render a component already opened, but it won't be updated anymore.
Any reaction to when a component is opened/closed has to be done
with the `onOpen`/`onClose` actions.