[FEATURE] Add `ember-basic-dropdown--focus-inside` class when something inside is focused.
When any element inside the dropdown gets the focus, the entire dropdown gets a
`ember-basic-dropdown--focus-inside` class. That includes elements gaining the
focus inside the trigger or inside the dropdown, even if the dropdown is attached
somewhere else in the DOM tree.

When some element inside looses the focus, that class is removed.

This class is only added when the element gaining the focus is **inside** the dropdown,
not when the element is the dropdown itself has the focus.

Note: The trigger is considered to be a sub-component of the dropdown, and therefore
when the trigger is focused the class is added.

This feature is implemented using the `focusin`/`focusout` events. When the dropdown
is rendered in the root of the body, the component adds those events to the component
itself and to the dropdown-content, so it works as expected. When the dropdown is redered
"in place", it is smart enough to not add the events to the dropdown-content.

There is no default styles added to this class.