Add/remove global events in the dropdown-content component
The reposition still happens in the parent component, but it events that
trigger ir are added/removed by the dropdown component using 
didInsertElement/willDestroyElement.

This is simpler because it removes the need of checking for the presence
of the dropdown, scheduling actions in the after-render queue or checking
for fast boot in many places.