[BREAKING] Calling e.preventDefault on the event no changes the component's behavior
After some experimentation, calling `e.preventDefault` was not a good idea. Calling it
on some events (keydown) by example will prevent other behaviour non related with this
component to work. And `return false` was already available.

Therefore, now return false is the new way to go to prevent the default behaviour
from happening.

The convention is:
- e.preventDefault will affect regular browser behaviour.
- return false will affect this component's behaviour only.