Add reproducible example of focus problem.
The reposition needs to be faster. Probably by not depending in ember data propagation.
At the end, the reposition method is already doing work outside ember (set width/height/left),
so I guess that it could do everything outside of ember, so it doesn't run into deprecations
because setting properties inside the didInsertElement