Make the first reposition apply styles directly instead of using ember.
This only affects the first reposition. The reason is that if the dropdown contains an input
with autofocus and it gets the focus before it has been property positioned, the page scrolls
to the top (or bottom, depending where the wormhole placeholder is) to put that input into
the viewport.
There is no way for any component inside the dropdown to wait to get the focus (do it in the
afterRender is not enough), therefore this hack.
The style property is still bound in ember, and successive repositions will go through the usual
ember binding system.