The `calculatePosition` method is now inside /utils so users can import it
This makes easier to the end user to build their own repositioning strategy that
reuses some of the default logic. By example, if a user likes the default strategy
but wants some extra padding, it is now possible to pass a `calculatePosition` that
in turns call the default one, and the programmes just has to increase the `styles.top`
property in the desired amount.