[BREAKING] Account for the containers X/Y position when calculating coordinates
The `calculatePosition` function now checks if the parent of the destination div has
position absolute or relative, and if so, substracts its X/Y coordinates to correct
the positioning of the content.
The breaking part is that now the signature of the function is:

`calculatePosition(trigger, content, destination, options)` instead of
`calculatePosition(trigger, content, options)`.

Closes #233