Rebase scrollbar fix (#438)
* Add config option to fix scrollbar clicking

As mentioned in #409, clicking a scrollbar will close an open basic-dropdown. This is due to mousedown being used instead of click. To not break existing behaviour, I've added a config option which can be passed to the parent container (basic-dropdown) or the content component.

* Address PR comments and change property name

This changes the approach from using a boolean flag for click type to passing in the click event type instead.
Additionally added comment to change to 'click' in 2.0

* Change tests to use rootEventType as well

Tests were still using the previous flag.

* Removed remaining useClickEvent flag

I missed this in an earlier pass.

* Add default rootEventType to onClose test

The default rootEventType is specified in the main component and passed down, it needs to be bound in this test for the test to pass.

* Reinstall dependencies