Update dependencies
There is an issue with Ember Engines caused by a bug in
`ember-element-helper@0.1.0`.
This issue is fixed in `ember-element-helper@0.2.0`.
Update this dependency along with other dependencies.

`ember-cli-htmlbars@4.0.8 makes `ember-cli-htmlbars-inline-precompile`
obsolete. Remove it and use new import path for hbs helper.

Other updates are minor and straightforward.

`ember-code-snippet` is not updated to ^3 because it contains breaking
changes. Leaving it on a previous version for now.