Use aria-controls instead of aria-owns for a relationship between the trigger and the content.
Closes #557.

aria-owns should not be used to make the content elements children of the button, since this is invalid accessibility structure. Screen readers like Orca on Linux don't handle this well.

Instead, use aria-controls to build a relationship. In a second step, add aria-owns to the parent of the trigger so the content gets appended as the last child of that container.